コアシステム監査に基づき実行時の安定性と共通基盤を強化
- 固定更新と入力保持を導入し、シミュレーションとUI更新を分離
- ルートモーションの周回処理を修正し、アニメーションイベントを実行系へ接続
- GPUバッファをフェンスに連動させ、使用中の領域への上書きを防止
- 衝突・接地、AABB、モデルの原点・高さ調整を共通化
- モデル・画像の非同期準備と資源共有、ロード失敗時の復帰処理を追加
- 非一様ボーンスケールの法線処理をCPU/GPUで統一
- FollowCamera、UTF-8パス変換、エラー診断を共通基盤へ集約
- ソース登録チェックとビルド・回帰・GPU・シェーダーの統合検証を追加

Debug/Releaseの本体ビルド、9種類の回帰テスト、
WARP GPUテストおよび全シェーダーの検証に成功。

+++ src/Framework/Animation/AnimationSampler.h
@@ -7,20 +7,22 @@
 class AnimationSampler
 {
 public:
+	// Takes the shared playback's local seconds. The exact duration samples the
+	// final key for root extraction; this sampler never wraps independently.
 	DirectX::XMVECTOR SampleTranslation(
 		const AnimationClip& clip,
 		const BoneAnimation& boneAnimation,
 		const BoneData& bindPose,
-		float animationTimeSeconds) const;
+		double animationTimeSeconds) const;
 
 	DirectX::XMMATRIX SampleLocalTransform(
 		const AnimationClip& clip,
 		const BoneAnimation& boneAnimation,
 		const BoneData& bindPose,
-		float animationTimeSeconds) const;
+		double animationTimeSeconds) const;
 
 private:
-	static double GetAnimationTimeTicks(const AnimationClip& clip, float animationTimeSeconds);
+	static double GetAnimationTimeTicks(const AnimationClip& clip, double animationTimeSeconds);
 	static float GetInterpolationAmount(double fromTime, double toTime, double animationTimeTicks);
 	static DirectX::XMVECTOR SampleVectorKey(
 		const std::vector<VectorAnimationKey>& keys,
